Retrieving nybbles from an identifier
In Pastry, identifiers are “thought of as a sequence of digits with base
2^b”.  Most of the literature, and the Java implementation, let you
choose different values for b at runtime.  For now, at least, we're
going to fix this at 4, which is always given as the “typical value”.
This means that we'll need to be able to extract out individual nybbles
from a Pastry identifier, since these are the digits in base 2^4=16.