Local node implementation
This patch adds an interface for the communication layer of a Pastry
node.  We've currently only implemented this interface for “local” nodes
— i.e., nodes that are all running within a single process on a single
machine.  For these nodes, sending a message simply requires moving the
binary message buffer into a queue of waiting messages.  (Nodes are also
grouped into a “context”, which technically allows you to have separate
networks of local nodes within a single process; this will be useful as
our test suite expands.)

Address information is stored in a “node reference”, which is also
subclassed for the different communication layers that we'll support.  A
local node is identified by a simple integer ID.