Node applications
This patch adds the public API for applications, which are the callbacks
that process the messages received by a Pastry node.  Each application
has a 32-bit identifier, and each node can only have one application
registered for any given application ID.  We'll have some standard
built-in applications, such as the one that maintains the nodes routing
table.