Routing tables
The `crs_routing_table` type will be used, surprisingly enough, to store
the routing table for a Pastry node.  We have some test cases that make
sure that we get the right routing table when we add a bunch of entries
to it, but nothing more complex than that just yet.