Node references
The `crs_node_ref` type stores a reference to a node in the Pastry
network.  This node may or may not be in the local process; this is the
type that will manage the actual communication with any remote nodes.
The existing `crs_node` type will only be used for nodes that are
definitely in the local process; this is the type that manages how the
local process responds to any messages that target the local node.