Context objects
We now store the list of local node objects in a “context”.  In addition
to managing the lookup of local nodes, the context is also responsible
for freeing those node objects.  In addition, each node reference is now
“owned” by the node that created it.  (We don't have the context object
manage the references, because we want local nodes to be consistent with
remote nodes, in that each node has its own distinct set of references.
In particular, this should allow us to test what happens when two nodes
with the same ID try to join the pastry network.)