Leaf sets
This patch adds a type for storing a Pastry' node's leaf set.  We
hard-code the size of the set (currently at 32 elements).  There isn't
currently a way to remove elements from the set; you can only add them.
There's also the (most important) function to find which entry of the
leaf set (if any) is closest to a node ID.