Leaf sets can contain each node twice
The `above` and `below` sets are not limited to the node references in
the half-ring above or below the local node.  Instead, they should
contain the first L/2 nodes that you can reach traveling in that
direction from the local node.  This means that if there are fewer than
L/2 nodes (that we know of) in the overlay network, then each of those
nodes should appear in both lists.