Only store nodes in leaf set once
Okay, I'm changing my mind on the patch a couple of links back, where we
followed the Java FreePastry implementation and allowed a node to appear
in a leaf set more than once.  (If there are fewer nodes on the overlay
network than the size of the leaf sef, then you'd encounter a particular
node in both directions.)  I can't see why that's useful, and it
complicates the leaf set maintenance.  Until we determine otherwise, I'm
going to go back to having each node appear in the leaf set at most
once.  As part of this change, we're now correctly calculating the
minimum and maximum node IDs in the leaf set.  (We had these fields
before, but weren't updating them.)