Calculate next routing hop
This patch adds a function that queries a node to see what it thinks the
next hop should be for a message sent to some destination Pastry ID,
according to the standard Pastry routing protocol.  This doesn't include
any mechanism yet for handling node failures; my hope is that we can
handle that by removing the offending node references, and then asking
for the next hop again.  (Java FreePastry uses iterators or somesuch
nonsense, but I think that might be overly complex for our design.)