Sending messages using the routing protocol
The new `crs_node_send_message` function uses the Pastry routing
protocol to send a message to some other node in the overlay network.
It will use the local leaf set and routing table to find out which node
is the next hop, and then send the message along to that next node.
The existing (internal) `crs_node_process_message` function does another
next-hop check, sending the message one step further along the routing
chain.  Eventually, a node will determine that the message should be
delivered locally, at which point it passes the message off to its
registered applications to deserialize and process the message.

To test this, we have a new helper application that just prints out any
messages that it receives, and a new `send` test command that can be
used to send a simple one-word message from one node to another.