Message parsing in public API
I've tweaked the existing internal helper functions for encoding and
decoding messages and put them into the public API.  They now operate on
the new `crs_message` type, which encapsulates the content of a message
sent across the overlay network.  Each message instance is either in
write-only mode or read-only mode.  Write-only mode allows you to append
data to the message, building up the message content in pieces.  You
then give the message to the overlay network, allowing it to route it to
its destination.  On the receiving end, we create a read-only
`crs_message` instance, which is decoded and passed to the application
that should process it.  During read-only mode, we maintain a cursor
into the message content, which is advanced forward as parts of the
message are decoded.  This lets us divide up the decoding responsibility
— the node itself is responsible for decoding the application ID, for
instance, while the application code is responsible for decoding
everything after that.

Right now when writing, all of the message content is concatenated into
a single buffer.  The API should allow us to replace that with some kind
of rope structure, if we want to try to support scatter/gather in the
I/O layer.