Construct node references from addresses
Before, the constructor function for a node reference took in both a
Pastry identifier and a network address that we should use to contact
the underlying node.  However, we won't always know in advance what
Pastry identifier to use.  In particular, for the bootstrap node, we
only have a network address.  So now it's the responsibility of the node
reference constructor to contact the node to see what it's Pastry
identifier is.