Applications can intercept messages
Each application can now provide an `intercept` callback, which is can
use to spy on or override the routing behavior of the overlay network.
The `intercept` callback is given a reference to the `next_hop` node for
the basic routing protocol.  The default implementation simply forwards
the message on to the `next_hop` node.  Some applications will want to
override this behavior, either to perform extra logic when a node sees a
message in transit, or to re-route the message in some
application-specific way.