Routing tables must contain local references
A “local” reference isn't a reference that refers to the local node,
it's one that's owned by the local node.  This is only really an issue
in our test cases — since everything is in the same process space, it's
easy to mix up the various `crs_node_ref` instances that refer to the
same node.  We now have some assert statements that should make sure
we don't let node references cross over into nodes that they don't
belong to.